Fix storage app db migration
Originally, running this migration caused the MySQL error (1215, 'Cannot add foreign key constraint') due to the deletion of a unique constraint on a field with a foreign key constraint. Removing the foreign key before and adding it back after deleting the unique constraint resolves the issue. See: http://stackoverflow.com/questions/8482346/mysql-cannot-drop-index-needed-in-a-foreign-key-constraint